pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "fwd-model-tools"
version = "0.1.0"
description = "Forward-modeling and sampling on top of JAXPM + JAX-Decomp"
authors = [
    { name = "Forward Model Tools Contributors" }
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["jax", "cosmology", "nbody", "sampling", "mcmc", "bayesian-inference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "jax>=0.6.0",
    "equinox",
    "blackjax",
    "numpyro",
    "jaxtyping",
    "jaxpm",
    "healpy",
    "jax-cosmo",
    "diffrax",
    "arviz",
    "getdist",
    "rich",
    "orbax-checkpoint",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
]
io = ["pyarrow>=14.0"]

[project.urls]
Homepage = "https://github.com/DifferentiableUniverseInitiative/fwd-model-tools"
Documentation = "https://fwd-model-tools.readthedocs.io"
Repository = "https://github.com/DifferentiableUniverseInitiative/fwd-model-tools"
Issues = "https://github.com/DifferentiableUniverseInitiative/fwd-model-tools/issues"


[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "UP",  # pyupgrade
    "T10", # flake8-debugger
    "I",   # isort
    "T20", # flake8-print
]

ignore = [
    "E501",  # line too long
    "E402",  # module level import not at top of file
    "E731",  # do not assign a lambda expression, use a def
    "F722", # syntax error in type annotation
    "UP045", # allow "Optional" usage instead of "X | None"
]

[tool.ruff.format]
quote-style = "double"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.yapf]
based_on_style = "pep8"
column_limit = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=fwd_model_tools --cov-report=term-missing"
